refactor: split the sidebar grouping and the session event subscriptions out
Two files sat past CLAUDE.md's 800-line max and this range pushed both
further. Each is cut along the seam it already had.

sessions.ts (874) held the session reducers plus how those sessions are
drawn in the sidebar. The grouping and the flat id orders a drag commits
back move to sidebar-groups.ts: sidebarGroups, runCardIn, reorderSubset,
dragOrder, neighborSessionId, groupByWorktree, orderGroups,
isLastWorktreeSession and the two stand-in group keys. sessions.ts is 668
lines, sidebar-groups.ts 232. The moved names are re-exported from
sessions.ts, so no caller and no test import changed.

projects.tsx (1095) held the provider plus twelve subscriptions to what
the backend announces about a session nobody in the window asked for.
Those move to project-events.tsx behind useSessionEvents, which takes the
refs and setters the effects already read. projects.tsx is 705 lines,
project-events.tsx 466.

Mechanical: no behaviour change, no renamed export, every dependency
array carried across as it was.
